+++ lib/shadcn_ui.ex
@@ -23,6 +23,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Overlays.Dialog,
     ShadcnUI.Components.Overlays.Drawer,
     ShadcnUI.Components.Overlays.Popover,
+    ShadcnUI.Components.Overlays.DropdownActions,
     ShadcnUI.Components.Foundation.Alert,
     ShadcnUI.Components.Foundation.Avatar,
     ShadcnUI.Components.Foundation.Badge,